HF_space_deployment/requirements.txt
tensorflow==2.16.1
keras==3.12.1
gradio==3.50.2
huggingface_hub==0.20.3
pillow
numpy
h5py
gunicorn


